Accept apiKey in constructor.
This simple change allows users to simply pass in their api key into the constructor.

Users might not have access to set environment variables (at the least, it can be inconvenient, especially in production), and passing the apiKey into the constructor is simpler and clearer than setting a system property.  In my case, it's easiest to get my apiKey via my play config, and then I don't have to set it twice. This allows me to call HoneyBadger easily inside of a play app, but outside the context of an http request (in an akka actor).

All the existing functionality should work exactly as-is, and is fully backwards compatible. This includes checking for the ENV variable, and the system property.