Make Causes special-case handling a little bit easier-to-read
The reason I even noticed this was a crashing bug with google's error-prone static analysis library.
This affected Java 12, which isn't officially supported by error-prone; we're also not quite ready to "support" Java 12,
at least when building from source, but the logic is hard to read, so I
figure a little simplification is in order.