Java 11-12, Spring 5 support (#111)
* Tweak travis to include JDK 10 and 11 in test matrix

* Add opendjk 10 and 11

* ErrorProne cleanup

Follows most of the ErrorProne advice.

* Make Causes special-case handling a little bit easier-to-read

The reason I even noticed this was a crashing bug with google's error-prone static analysis library.
This affected Java 12, which isn't officially supported by error-prone; we're also not quite ready to "support" Java 12,
at least when building from source, but the logic is hard to read, so I
figure a little simplification is in order.

* Update dependencies

Uses newer Maven plugins, tweaks Error-Prone config to support building
with Java 1.8 as well as more modern ones (11).

Todo: Maybe support Java 12 (but this requires disabling google's error-prone for now, due
to crashing defects in the static analysis that error-prone does).

* Tentatively adding Java 12 to the build matrix

(Do we feel lucky? Works on my machine!)

* Update Readme to reflect current supported versions

I need to do a little bit more testing and probably refresh crywolf-java
before I call it good, but this looks like the bulk of the work.